dia: run_to_sample_codes lives in plots.general to avoid an import cycle
plots.dia imported it from dia_utils, and dia_utils imports plots.dia.
MultiQC loads dia_utils first, so the partially initialised module made
the import fail and the whole QuantMS module was skipped - the report was
written without it. A test now imports dia_utils first in a fresh
interpreter.
