+++ pmultiqc/modules/common/dia_utils.py
@@ -1,5 +1,6 @@
 import itertools
 import numpy as np
+from pmultiqc.modules.common.plots.general import run_to_sample_codes
 import pandas as pd
 import re
 from collections import OrderedDict
@@ -315,25 +316,6 @@ def _process_run_data(df, ms_with_psm, quantms_modified, sdrf_file_df):
     )
 
     return cal_num_table_data
-
-
-def run_to_sample_codes(runs: pd.Series, file_df: pd.DataFrame) -> pd.Series:
-    """Sample id per row from the Run column, without materialising strings.
-
-    Merging the report with the SDRF sample table on ``Run`` was measured to be
-    the largest transient in the summary: it upcasts the categorical key to
-    object for every row (231 M on PXD030304) and adds an object Sample column.
-    Mapping through the category codes touches one small array instead.
-    Rows whose run is not in ``file_df`` get NaN.
-    """
-    run_to_sample = file_df[["Run", "Sample"]].drop_duplicates().set_index("Run")["Sample"]
-    run_to_sample.index = run_to_sample.index.astype(str)
-    if isinstance(runs.dtype, pd.CategoricalDtype):
-        per_category = run_to_sample.reindex(runs.cat.categories.astype(str)).to_numpy(dtype="float64", na_value=np.nan)
-        codes = runs.cat.codes.to_numpy()
-        values = np.where(codes >= 0, per_category[np.clip(codes, 0, None)], np.nan)
-        return pd.Series(values, index=runs.index)
-    return runs.astype(str).map(run_to_sample).astype("float64")
 
 
 def _is_modified(sequences: pd.Series) -> pd.Series:
