+++ pmultiqc/modules/common/plots/general.py
@@ -505,3 +505,22 @@ def summarise_box_data(plot_data, whisker_iqr=1.5, max_points=None):
         summarised.append(out)
 
     return summarised if was_list else summarised[0]
+
+
+def run_to_sample_codes(runs: pd.Series, file_df: pd.DataFrame) -> pd.Series:
+    """Sample id per row from the Run column, without materialising strings.
+
+    Merging the report with the SDRF sample table on ``Run`` was measured to be
+    the largest transient in the summary: it upcasts the categorical key to
+    object for every row (231 M on PXD030304) and adds an object Sample column.
+    Mapping through the category codes touches one small array instead.
+    Rows whose run is not in ``file_df`` get NaN.
+    """
+    run_to_sample = file_df[["Run", "Sample"]].drop_duplicates().set_index("Run")["Sample"]
+    run_to_sample.index = run_to_sample.index.astype(str)
+    if isinstance(runs.dtype, pd.CategoricalDtype):
+        per_category = run_to_sample.reindex(runs.cat.categories.astype(str)).to_numpy(dtype="float64", na_value=np.nan)
+        codes = runs.cat.codes.to_numpy()
+        values = np.where(codes >= 0, per_category[np.clip(codes, 0, None)], np.nan)
+        return pd.Series(values, index=runs.index)
+    return runs.astype(str).map(run_to_sample).astype("float64")
